DAOS-4605 vos: Add tool to estimate SCM ratio
Add the vos_scm_estimator.py tool. It merges the functionality
of the former tools vos_size, vos_size.py, vos_size_dfs_sample.py
and parse_csv.py. It removes the need of running individual
tools in a pipeline to get an estimation if the SCM.

Vos and DFS layouts are retrieved from the DAOS libraries. It
is possible to estimate the SCM/NVM ratio from csv files,
yaml files and even from the file system.

Skip-func-hw-test: true

Signed-off-by: Martinez Montes, Jonathan <jonathan.martinez.montes@intel.com>